refactor(storefront/orpc): move the file under `shared/api`
