+++ apps/storefront/src/app/[lang]/page.tsx
@@ -10,7 +10,7 @@ import { ComingSoon } from "@/components/app/coming-soon";
 import { HeroBanner } from "@/components/app/hero-banner";
 import { ProductGrid } from "@/components/app/product-grid";
 import { ProductSectionHeader } from "@/components/app/product-section-header";
-import { client, getQueryClient, orpc } from "@/lib/orpc";
+import { client, getQueryClient, orpc } from "@/shared/api/orpc";
 import { getStoreSlug } from "@/shared/lib/store/slug-retrieval.util";
 
 export default async function StorePage() {
